Замена библиотеки FFTW на более простой алгоритм
- FFTW начинает работать существенно медленнее при обработке массивов большого количества одинаковых данных

* передача данных из класса cy3device - по указателю на qvector(char), а не объектом qvector(short)
* коррекция расчёта пропускной способности (теперь - в double с меньшей вероятностью округления/переполнения)

fix: ошибка размерности массива окна БПФ
fix: ошибка инициализации объекта USBDevice

+++ mainwindow.h
@@ -26,7 +26,7 @@ public:
 private slots:
     void on_buttonOpenDevice_clicked();
     void DebugParser(QString Message);
-    void DisplayBandwidth(int BW);
+    void DisplayBandwidth(double BW);
 
     void handleAbortDump();
